kg/pyproject.toml
[project]
name = "kg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.33.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-ollama>=0.3.7",
    "neo4j>=5.28.2",
    "python-dotenv>=1.1.1",
]
